Switch Unity 6 entity IDs to signed long

+++ Assets/DONT TOUCH/Scripts/Extensions/EntityIdExtensions.cs
@@ -5,14 +5,14 @@ namespace DONT_TOUCH.Scripts.Extensions
     public static class EntityIdExtensions
     {
 #if UNITY_6000_5_OR_NEWER
-        public static ulong GetId(this EntityId entityId)
+        public static long GetId(this EntityId entityId)
         {
-            return EntityId.ToULong(entityId);
+            return (long)EntityId.ToULong(entityId);
         }
 
-        public static ulong GetId(this Object target)
+        public static long GetId(this Object target)
         {
-            return EntityId.ToULong(target.GetEntityId());
+            return (long)EntityId.ToULong(target.GetEntityId());
         }
 #else
         public static int GetId(this Object target)
